oz-backend-django/pyproject.toml
[tool.poetry]
name = "oz-backend-django"
version = "0.1.0"
description = ""
authors = ["tanmtn <tksdl8312@naver.com>"]
readme = "README.md"
packages = [{include = "oz_backend_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
